+++ spigot-nms/src/main/java/com/kamikazejam/kamicommon/nms/provider/MessageManagerProvider.java
@@ -13,7 +13,7 @@ import org.jetbrains.annotations.NotNull;
  */
 public class MessageManagerProvider extends Provider<AbstractMessageManager> {
     @Override
-    protected @NotNull AbstractMessageManager provide(double ver) {
+    protected @NotNull AbstractMessageManager provide(int ver) {
         if (ver < f("1.8")) {
             throw new IllegalArgumentException("Version not supported (< 1.8): " + ver);
         }
